Added Picture Support, and a way to change .exe in the case of PlayAction.Type being a URL.
Background picture, hero photo, and misc photo support in Steam. Grabs from game's background and cover photos.
Made it so users are prompted on .exe file if the current PlayAction is a URL.
Made it so that Steam AppId is the preferred naming schema for support with Steam Controller. (Going to change later to default to name, steam appid with a setting somewhere).